fix: correct color of piece and rotation

+++ components/container.go
@@ -1,13 +1,13 @@
 package components
 
 type Container struct {
-	OriginX   int
-	OriginY   int
+	OriginX   float64
+	OriginY   float64
 	Children  []Block
 	Transform Transform
 }
 
-func (container *Container) GetChildAbsPosition(block *Block) (int, int) {
+func (container *Container) GetChildAbsPosition(block *Block) (float64, float64) {
 	x, y := container.OriginX, container.OriginY
 	dx, dy := container.Transform.TranformPosition(block.OriginX, block.OriginY)
 	return x + dx, y + dy
